pyproject.toml
[tool.poetry]
authors = ["Justine Kizhakkinedath <justine.kizhak@mindstix.com>"]
description = ""
name = "python-baseline"
version = "0.1.0"

[tool.poetry.dependencies]
fastapi = "^0.62.0"
python = "^3.8"
uvicorn = "^0.12.3"
python-dotenv = "^0.15.0"
typer = "^0.3.2"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]
black = {version = "^19.10b0", allow-prereleases = true}
flake8 = "^3.8.3"
mypy = "^0.780"
pdbpp = "^0.10.2"
pylint = "^2.5.3"
pytest = "^5.4.3"
pytest-cov = "^2.10.0"
pytest-watch = "^4.2.0"
isort = "^5.6.4"
autoflake = "^1.4"

[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]
